// generated/io-fn-app/NewProfile.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { EmailAddress } from "./EmailAddress";
import { IsEmailValidated } from "./IsEmailValidated";
import { IsTestProfile } from "./IsTestProfile";
import * as t from "io-ts";

/**
 * Describes a new citizen's profile.
 * Used for profile creation.
 */

// required attributes
const NewProfileR = t.interface({
  is_email_validated: IsEmailValidated
});

// optional attributes
const NewProfileO = t.partial({
  email: EmailAddress,

  is_test_profile: IsTestProfile
});

export const NewProfile = t.intersection(
  [NewProfileR, NewProfileO],
  "NewProfile"
);

export type NewProfile = t.TypeOf<typeof NewProfile>;