// generated/io-fn-app/Profile.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { EmailAddress } from "./EmailAddress";
import { BlockedInboxOrChannels } from "./BlockedInboxOrChannels";
import { PreferredLanguages } from "./PreferredLanguages";
import { IsInboxEnabled } from "./IsInboxEnabled";
import { AcceptedTosVersion } from "./AcceptedTosVersion";
import { IsWebhookEnabled } from "./IsWebhookEnabled";
import { IsEmailEnabled } from "./IsEmailEnabled";
import * as t from "io-ts";

/**
 * Describes the citizen's profile.
 * Used for profile update.
 */

// required attributes
const ProfileR = t.interface({
  version: t.Integer
});

// optional attributes
const ProfileO = t.partial({
  email: EmailAddress,

  blocked_inbox_or_channels: BlockedInboxOrChannels,

  preferred_languages: PreferredLanguages,

  is_inbox_enabled: IsInboxEnabled,

  accepted_tos_version: AcceptedTosVersion,

  is_webhook_enabled: IsWebhookEnabled,

  is_email_enabled: IsEmailEnabled
});

export const Profile = t.intersection([ProfileR, ProfileO], "Profile");

export type Profile = t.TypeOf<typeof Profile>;